[BEAM-6186] Optimization cleanup: move phase utilities out of local scope.
