+++ sdks/python/apache_beam/runners/portability/fn_api_runner.py
@@ -52,6 +52,9 @@ from apache_beam.portability.api import beam_fn_api_pb2_grpc
 from apache_beam.portability.api import beam_runner_api_pb2
 from apache_beam.portability.api import endpoints_pb2
 from apache_beam.runners import pipeline_context
+from apache_beam.runners.portability import fn_api_runner_transforms
+from apache_beam.runners.portability.fn_api_runner_transforms import only_element
+from apache_beam.runners.portability.fn_api_runner_transforms import unique_name
 from apache_beam.runners import runner
 from apache_beam.runners.worker import bundle_processor
 from apache_beam.runners.worker import data_plane
@@ -292,101 +295,15 @@ class FnApiRunner(runner.PipelineRunner):
 
   def create_stages(self, pipeline_proto):
 
-    # First define a couple of helpers.
-
-    def union(a, b):
-      # Minimize the number of distinct sets.
-      if not a or a == b:
-        return b
-      elif not b:
-        return a
-      else:
-        return frozenset.union(a, b)
-
-    class Stage(object):
-      """A set of Transforms that can be sent to the worker for processing."""
-      def __init__(self, name, transforms,
-                   downstream_side_inputs=None, must_follow=frozenset()):
-        self.name = name
-        self.transforms = transforms
-        self.downstream_side_inputs = downstream_side_inputs
-        self.must_follow = must_follow
-        self.timer_pcollections = []
-
-      def __repr__(self):
-        must_follow = ', '.join(prev.name for prev in self.must_follow)
-        if self.downstream_side_inputs is None:
-          downstream_side_inputs = '<unknown>'
-        else:
-          downstream_side_inputs = ', '.join(
-              str(si) for si in self.downstream_side_inputs)
-        return "%s\n  %s\n  must follow: %s\n  downstream_side_inputs: %s" % (
-            self.name,
-            '\n'.join(["%s:%s" % (transform.unique_name, transform.spec.urn)
-                       for transform in self.transforms]),
-            must_follow,
-            downstream_side_inputs)
-
-      def can_fuse(self, consumer):
-        def no_overlap(a, b):
-          return not a.intersection(b)
-        return (
-            not self in consumer.must_follow
-            and not self.is_flatten() and not consumer.is_flatten()
-            and no_overlap(self.downstream_side_inputs, consumer.side_inputs()))
-
-      def fuse(self, other):
-        return Stage(
-            "(%s)+(%s)" % (self.name, other.name),
-            self.transforms + other.transforms,
-            union(self.downstream_side_inputs, other.downstream_side_inputs),
-            union(self.must_follow, other.must_follow))
-
-      def is_flatten(self):
-        return any(transform.spec.urn == common_urns.primitives.FLATTEN.urn
-                   for transform in self.transforms)
-
-      def side_inputs(self):
-        for transform in self.transforms:
-          if transform.spec.urn == common_urns.primitives.PAR_DO.urn:
-            payload = proto_utils.parse_Bytes(
-                transform.spec.payload, beam_runner_api_pb2.ParDoPayload)
-            for side_input in payload.side_inputs:
-              yield transform.inputs[side_input]
-
-      def has_as_main_input(self, pcoll):
-        for transform in self.transforms:
-          if transform.spec.urn == common_urns.primitives.PAR_DO.urn:
-            payload = proto_utils.parse_Bytes(
-                transform.spec.payload, beam_runner_api_pb2.ParDoPayload)
-            local_side_inputs = payload.side_inputs
-          else:
-            local_side_inputs = {}
-          for local_id, pipeline_id in transform.inputs.items():
-            if pcoll == pipeline_id and local_id not in local_side_inputs:
-              return True
-
-      def deduplicate_read(self):
-        seen_pcolls = set()
-        new_transforms = []
-        for transform in self.transforms:
-          if transform.spec.urn == bundle_processor.DATA_INPUT_URN:
-            pcoll = only_element(list(transform.outputs.items()))[1]
-            if pcoll in seen_pcolls:
-              continue
-            seen_pcolls.add(pcoll)
-          new_transforms.append(transform)
-        self.transforms = new_transforms
-
     # Some helper functions.
+    # TODO(BEAM-6186): Move these to fn_api_runner_transforms.
+
+    Stage = fn_api_runner_transforms.Stage
+    union = fn_api_runner_transforms.union
 
     def add_or_get_coder_id(coder_proto):
-      for coder_id, coder in pipeline_components.coders.items():
-        if coder == coder_proto:
-          return coder_id
-      new_coder_id = unique_name(pipeline_components.coders, 'coder')
-      pipeline_components.coders[new_coder_id].CopyFrom(coder_proto)
-      return new_coder_id
+      return fn_api_runner_transforms.TransformContext(
+          pipeline_components).add_or_get_coder_id(coder_proto)
 
     def windowed_coder_id(coder_id, window_coder_id):
       proto = beam_runner_api_pb2.Coder(
@@ -1006,27 +923,10 @@ class FnApiRunner(runner.PipelineRunner):
 
     pipeline_components = copy.deepcopy(pipeline_proto.components)
 
-    known_composites = set(
-        [common_urns.primitives.GROUP_BY_KEY.urn,
-         common_urns.composites.COMBINE_PER_KEY.urn])
-
-    def leaf_transforms(root_ids):
-      for root_id in root_ids:
-        root = pipeline_proto.components.transforms[root_id]
-        if root.spec.urn in known_composites:
-          yield root_id
-        elif not root.subtransforms:
-          # Make sure its outputs are not a subset of its inputs.
-          if set(root.outputs.values()) - set(root.inputs.values()):
-            yield root_id
-        else:
-          for leaf in leaf_transforms(root.subtransforms):
-            yield leaf
-
     # Initial set of stages are singleton leaf transforms.
-    stages = [
-        Stage(name, [pipeline_proto.components.transforms[name]])
-        for name in leaf_transforms(pipeline_proto.root_transform_ids)]
+    stages = list(fn_api_runner_transforms.leaf_transform_stages(
+        pipeline_proto.root_transform_ids,
+        pipeline_proto.components))
 
     # Apply each phase in order.
     for phase in [
@@ -1658,23 +1558,6 @@ class RunnerResult(runner.PipelineResult):
       self._monitoring_metrics = FnApiMetrics(
           self._monitoring_infos_by_stage, user_metrics_only=False)
     return self._monitoring_metrics
-
-
-def only_element(iterable):
-  element, = iterable
-  return element
-
-
-def unique_name(existing, prefix):
-  if prefix in existing:
-    counter = 0
-    while True:
-      counter += 1
-      prefix_counter = prefix + "_%s" % counter
-      if prefix_counter not in existing:
-        return prefix_counter
-  else:
-    return prefix
 
 
 def create_buffer_id(name, kind='materialize'):
